move context menu

+++ Bar.xaml.cs
@@ -63,8 +63,8 @@ public partial class Bar : Window
 		this.Background = Utils.BrushFromHex(config.backgroundColor);
 		if(this.Background.Equals(Colors.Transparent)) { barTransparent = true; }
 
-		int exStyles = Win32.GetWindowLong(hWnd, -20);
-		Win32.SetWindowLong(hWnd, -20, exStyles | (int)WindowStyles.WS_EX_TOOLWINDOW);
+		uint exStyles = Win32.GetWindowLong(hWnd, -20);
+		Win32.SetWindowLong(hWnd, -20, (int)(exStyles | (uint)WindowStyles.WS_EX_TOOLWINDOW));
 
 		//Win32.SetWindowPos(hWnd, IntPtr.Zero, config.marginXLeft, config.marginYTop, config.width, config.height, 0x0400);
 		this.Width = config.width;
